requirements.txt
pymupdf
openai
instructor
ipykernel
tiktoken
pandas
python-dotenv
cohere
numpy
fastapi
uvicorn
PyPDF2
firebase-admin
python-multipart
requests 